refactor: template, temporal memo archive 로직 위치 변경 & url 수정
