+++ src/main/java/com/baro/memo/application/TemporalMemoService.java
@@ -1,9 +1,13 @@
 package com.baro.memo.application;
 
 
+import com.baro.archive.domain.Archive;
+import com.baro.archive.domain.ArchiveRepository;
 import com.baro.member.domain.Member;
 import com.baro.member.domain.MemberRepository;
 import com.baro.memo.application.dto.ApplyCorrectionCommand;
+import com.baro.memo.application.dto.ArchiveTemporalMemoCommand;
+import com.baro.memo.application.dto.ArchiveTemporalMemoResult;
 import com.baro.memo.application.dto.DeleteTemporalMemoCommand;
 import com.baro.memo.application.dto.FindTemporalMemoHistoriesQuery;
 import com.baro.memo.application.dto.FindTemporalMemoHistoriesResult;
@@ -16,6 +20,8 @@ import com.baro.memo.domain.TemporalMemo;
 import com.baro.memo.domain.TemporalMemoRepository;
 import com.baro.memo.exception.TemporalMemoException;
 import com.baro.memo.exception.TemporalMemoExceptionType;
+import com.baro.memofolder.domain.MemoFolder;
+import com.baro.memofolder.domain.MemoFolderRepository;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Comparator;
@@ -33,6 +39,8 @@ public class TemporalMemoService {
 
     private final TemporalMemoRepository temporalMemoRepository;
     private final MemberRepository memberRepository;
+    private final MemoFolderRepository memoFolderRepository;
+    private final ArchiveRepository archiveRepository;
 
     public SaveTemporalMemoResult saveTemporalMemo(SaveTemporalMemoCommand command) {
         Member member = memberRepository.getById(command.memberId());
@@ -90,5 +98,17 @@ public class TemporalMemoService {
                 .map(FindTemporalMemoResult::from)
                 .sorted(Comparator.comparing(FindTemporalMemoResult::createdAt))
                 .toList();
+    }
+
+    public ArchiveTemporalMemoResult archiveTemporalMemo(ArchiveTemporalMemoCommand command) {
+        Member member = memberRepository.getById(command.memberId());
+        TemporalMemo temporalMemo = temporalMemoRepository.getById(command.temporalMemoId());
+        temporalMemo.matchOwner(member.getId());
+        MemoFolder memoFolder = memoFolderRepository.getById(command.memoFolderId());
+        memoFolder.matchOwner(member.getId());
+
+        Archive archive = archiveRepository.save(new Archive(member, memoFolder, temporalMemo.getContent()));
+        temporalMemo.archived(archive);
+        return ArchiveTemporalMemoResult.from(archive);
     }
 }
