+++ src/main/java/com/baro/template/presentation/TemplateController.java
@@ -55,4 +55,32 @@ public class TemplateController {
         templateService.copyTemplate(command);
         return ResponseEntity.noContent().build();
     }
+
+
+    @PostMapping("/{templateId}/archive")
+    ResponseEntity<Void> archiveTemplate(
+            AuthMember authMember,
+            @RequestBody ArchiveTemplateRequest request,
+            @PathVariable("templateId") Long templateId
+    ) {
+        ArchiveTemplateCommand command = new ArchiveTemplateCommand(authMember.id(), templateId,
+                request.memoFolderId());
+        ArchiveTemplateResult result = templateService.archiveTemplate(command);
+
+        URI location = ServletUriComponentsBuilder.fromPath("/archives/templates")
+                .path("/{id}")
+                .buildAndExpand(result.id())
+                .toUri();
+        return ResponseEntity.created(location).build();
+    }
+
+    @DeleteMapping("/{templateId}/archive")
+    ResponseEntity<Void> unarchiveTemplate(
+            AuthMember authMember,
+            @PathVariable Long templateId
+    ) {
+        UnArchiveTemplateCommand command = new UnArchiveTemplateCommand(authMember.id(), templateId);
+        templateService.unarchiveTemplate(command);
+        return ResponseEntity.noContent().build();
+    }
 }
