+++ src/test/java/com/baro/common/acceptance/template/TemplateAcceptanceSteps.java
@@ -197,4 +197,95 @@ public class TemplateAcceptanceSteps {
                 fieldWithPath("content[].copiedCount").description("템플릿 복사 횟수")
         );
     }
+
+    public static ExtractableResponse<Response> 템플릿_아카이브_요청_성공(Token 토큰, Long 템플릿_ID, ArchiveTemplateRequest 바디) {
+        var url = "/templates/{templateId}/archive";
+
+        return given(requestSpec).log().all()
+                .filter(document(DEFAULT_REST_DOCS_PATH,
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")
+                        ),
+                        pathParameters(
+                                parameterWithName("templateId").description("템플릿 ID")
+                        ),
+                        requestFields(
+                                fieldWithPath("memoFolderId").description("메모 폴더 ID")
+                        ),
+                        responseHeaders(
+                                headerWithName(HttpHeaders.LOCATION).description("아카이브된 템플릿 경로")
+                        )
+                ))
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + 토큰.accessToken()).body(바디)
+                .pathParam("templateId", 템플릿_ID)
+                .when().post(url)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 템플릿_아카이브_요청_실패(Token 토큰, Long 템플릿_ID, ArchiveTemplateRequest 바디) {
+        var url = "/templates/{templateId}/archive";
+
+        return given(requestSpec).log().all()
+                .filter(document(DEFAULT_REST_DOCS_PATH,
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")
+                        ),
+                        pathParameters(
+                                parameterWithName("templateId").description("템플릿 ID")
+                        ),
+                        requestFields(
+                                fieldWithPath("memoFolderId").description("메모 폴더 ID")
+                        ),
+                        responseFields(예외_응답())
+                ))
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + 토큰.accessToken()).body(바디)
+                .pathParam("templateId", 템플릿_ID)
+                .when().post(url)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 템플릿_아카이브_취소_요청_성공(Token 토큰, Long 템플릿_ID) {
+        var url = "/templates/{templateId}/archive";
+
+        return given(requestSpec).log().all()
+                .filter(document(DEFAULT_REST_DOCS_PATH,
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")
+                        ),
+                        pathParameters(
+                                parameterWithName("templateId").description("템플릿 ID")
+                        )
+                ))
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + 토큰.accessToken())
+                .pathParam("templateId", 템플릿_ID)
+                .when().delete(url)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 템플릿_아카이브_취소_요청_실패(Token 토큰, Long 템플릿_ID) {
+        var url = "/templates/{templateId}/archive";
+
+        return given(requestSpec).log().all()
+                .filter(document(DEFAULT_REST_DOCS_PATH,
+                        requestHeaders(
+                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 토큰")
+                        ),
+                        pathParameters(
+                                parameterWithName("templateId").description("템플릿 ID")
+                        ),
+                        responseFields(예외_응답())
+                ))
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + 토큰.accessToken())
+                .pathParam("templateId", 템플릿_ID)
+                .when().delete(url)
+                .then().log().all()
+                .extract();
+    }
 }
